Modal close icons edited
