+++ assets/client/components/TaskBar.tsx
@@ -12,15 +12,10 @@ import { localizer } from '../Localizer';
 
 export class TaskBar extends VitrineComponent {
 	private currentWindow: BrowserWindow;
-	private colorStyles: any;
 
 	public constructor(props: any) {
 		super(props);
 		this.currentWindow = remote.getCurrentWindow();
-		this.colorStyles = {
-			color: rgba(255, 255, 255, 0.3),
-			hover: rgba(255, 255, 255, 0.6)
-		};
 	}
 
 	private minimizeBtnClickHandler() {
@@ -94,18 +89,12 @@ export class TaskBar extends VitrineComponent {
 				{taskBarElements}
 				<div className={css(styles.windowControlBtnGroup)}>
 					<MinimizeIcon
-						styles={styles.windowControlIcon}
-						colors={this.colorStyles}
 						onClick={this.minimizeBtnClickHandler.bind(this)}
 					/>
 					<MaximizeIcon
-						styles={styles.windowControlIcon}
-						colors={this.colorStyles}
 						onClick={this.maximizeBtnClickHandler.bind(this)}
 					/>
 					<CloseIcon
-						styles={styles.windowControlIcon}
-						colors={this.colorStyles}
 						onClick={this.closeBtnClickHandler.bind(this)}
 					/>
 				</div>
@@ -144,10 +133,5 @@ const styles: React.CSSProperties = StyleSheet.create({
 		top: 16,
 		right: 20,
 		zIndex: 1
-	},
-	windowControlIcon: {
-		cursor: 'pointer',
-		marginLeft: 14,
-		'-webkitAppRegion': 'no-drag'
 	}
 });
