+++ assets/client/components/icons/MinimizeIcon.tsx
@@ -1,25 +1,32 @@
 import * as React from 'react';
 import { StyleSheet, css } from 'aphrodite';
+import { rgba } from 'css-verbose';
 
 import { VitrineComponent } from '../VitrineComponent';
 
 export class MinimizeIcon extends VitrineComponent {
+	private colorStyles: any;
+
 	public constructor(props: any) {
 		super(props);
+		this.colorStyles = {
+			color: rgba(255, 255, 255, 0.3),
+			hover: rgba(255, 255, 255, 0.6)
+		};
 		this.state = {
-			color: this.props.colors.color
+			color: this.colorStyles.color
 		};
 	}
 
 	private mouseEnterHandler() {
 		this.setState({
-			color: this.props.colors.hover
+			color: this.colorStyles.hover
 		});
 	}
 
 	private mouseLeaveHandler() {
 		this.setState({
-			color: this.props.colors.color
+			color: this.colorStyles.color
 		});
 	}
 
@@ -28,7 +35,7 @@ export class MinimizeIcon extends VitrineComponent {
 			<svg
 				width={16}
 				height={16}
-				className={css(this.props.styles)}
+				className={css(styles.windowControlIcon)}
 				onClick={this.props.onClick}
 				onMouseEnter={this.mouseEnterHandler.bind(this)}
 				onMouseLeave={this.mouseLeaveHandler.bind(this)}
@@ -50,5 +57,10 @@ export class MinimizeIcon extends VitrineComponent {
 const styles: React.CSSProperties = StyleSheet.create({
 	btn: {
 		transition: `${200}ms ease`
+	},
+	windowControlIcon: {
+		cursor: 'pointer',
+		marginLeft: 14,
+		'-webkitAppRegion': 'no-drag'
 	}
 });
